remove old docs and refactory new structure
