--- docs/env.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="vite/client" />
-declare module 'vue-techs-logos';
-declare module 'darkmode-vue';
