--- docs/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "docs",
-  "version": "0.1.5",
-  "private": true,
-  "scripts": {
-    "dev": "vite",
-    "build": "run-p type-check build-only",
-    "preview": "vite preview",
-    "test:unit": "vitest",
-    "test:e2e": "start-server-and-test preview http://localhost:4173 'cypress run --e2e'",
-    "test:e2e:dev": "start-server-and-test 'vite dev --port 4173' http://localhost:4173 'cypress open --e2e'",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
-  },
-  "dependencies": {
-    "darkmode-vue": "^0.4.0",
-    "sass": "^1.77.6",
-    "typescritp": "^1.0.0",
-    "vue": "^3.4.31",
-    "vue-techs-logos": "^0.1.15"
-  },
-  "devDependencies": {
-    "@rushstack/eslint-patch": "^1.10.3",
-    "@types/jsdom": "^21.1.7",
-    "@types/node": "^20.14.9",
-    "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/eslint-config-typescript": "^13.0.0",
-    "@vue/test-utils": "^2.4.6",
-    "@vue/tsconfig": "^0.5.1",
-    "cypress": "^13.12.0",
-    "eslint": "^9.6.0",
-    "eslint-plugin-cypress": "^3.3.0",
-    "eslint-plugin-vue": "^9.26.0",
-    "jsdom": "^24.1.0",
-    "npm-run-all": "^4.1.5",
-    "start-server-and-test": "^2.0.4",
-    "typescript": "^5.5.2",
-    "vite": "^5.3.2",
-    "vitest": "^1.6.0",
-    "vue-tsc": "^2.0.22"
-  }
-}
