--- docs/src/main.ts
@@ -1,6 +1,0 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-
-import './assets/main.css'
-
-createApp(App).mount('#app')
