+++ tsconfig.node.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.json",
-  "include": ["vite.config.*", "vitest.config.*", "playwright.config.*"],
+  "include": ["vite.config.*", "vitest.config.*", "playwright.config.*", "vite-module.config.ts"],
   "compilerOptions": {
     "composite": true,
     "types": ["node"],
